requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
annotated-types==0.7.0
anthropic==0.32.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.1.1
courlan==1.2.0
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
Faker==26.0.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
favicon==0.7.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
flatbuffers==23.5.26
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google_search_results==2.4.2
googleapis-common-protos==1.63.2
graphlib_backport==1.1.0
greenlet==3.0.3
grpcio==1.65.2
h11==0.14.0
htbuilder==0.6.2
html2text==2024.2.26
htmldate==1.8.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
humanize==4.10.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
jusText==3.0.1
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.10
langchain-anthropic==0.1.22
langchain-community==0.2.9
langchain-core==0.2.27
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langsmith==0.1.93
llama-cloud==0.0.9
llama-index==0.10.56
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.12
llama-index-core==0.10.56
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.26
llama-index-multi-modal-llms-openai==0.1.8
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-index-readers-web==0.1.23
llama-parse==0.4.9
llama_cpp_python==0.2.84
lxml==5.2.2
lxml_html_clean==0.1.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
merkle-json==1.0.0
millify==0.1.1
minijinja==2.0.1
mistune==3.0.2
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.15.1
openai==1.36.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
playwright==1.45.1
posthog==3.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
pypdf==4.3.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
selenium==4.23.0
Send2Trash==1.8.3
sentence-transformers==3.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spider-client==0.0.27
SQLAlchemy==2.0.31
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
streamlit-aggrid==0.3.4
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-pills==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
striprtf==0.0.26
sympy==1.12
tbb==2021.13.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tinysegmenter==0.3
tld==0.13
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
trafilatura==1.11.0
traitlets==5.14.3
transformers==4.42.4
trio==0.26.0
trio-websocket==0.11.1
trulens_eval==0.33.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.4
validators==0.33.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-search==2.1.2
zipp==3.19.2
